pyproject.toml
[tool.poetry]
name = "swa"
version = "0.1.0"
description = "Software for Halloween project 2024"
authors = ["cahogan <cahogan16@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
pytest-django = "^4.9.0"
pytest = "^8.3.3"
pytest-env = "^1.1.5"
psycopg = "^3.2.3"
pillow = "^11.0.0"
zebrafy = "^1.2.1"
pdf417 = "^0.8.1"

[tool.pytest.ini_options]
minversion = "6.0"
python_files = "test_*.py"
DJANGO_SETTINGS_MODULE = "swa.settings"

[tool.pytest_env]
DJANGO_SECRET_KEY = 'insecure-testing-only-key'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
